added validate spot data into utils/validation
