+++ backend/db/seeders/20240328040714-demo-spot-image.js
@@ -40,6 +40,16 @@ module.exports = {
         preview: true,
         spotId: 5
       },
+      {
+        url: 'image url 5',
+        preview: true,
+        spotId: 6
+      },
+      {
+        url: 'image url 5',
+        preview: true,
+        spotId: 7
+      },
     ])
   },
